requirements.txt
argcomplete==1.8.2
beautifulsoup4==4.5.3
chardet==2.3.0
Django==2.1.2
django-elasticsearch-dsl==0.5.0
django-extensions==2.1.3
django-rest-framework==0.1.0
djangorestframework==3.9.0
docx2txt==0.6
EbookLib==0.15
elasticsearch==6.3.1
elasticsearch-dsl==6.2.1
ipaddress==1.0.22
lxml==4.2.5
Pillow==5.3.0
pocketsphinx==0.1.3
python-dateutil==2.7.3
python-pptx==0.6.5
pytz==2018.5
six==1.10.0
SpeechRecognition==3.6.3
textract==1.6.1
urllib3==1.24
xlrd==1.0.0
XlsxWriter==1.1.2
